Remove explicit version from composer.json
Composer and Packagist determine the version by looking at the GIT tags.
"version" should only ever be used if there is no other way to determine
the version.

It creates ambiguity, as multiple commits are of the same version - without
the tag it is impossible to know if the first, the last or any commit
inbetween is the "correct" one for the version.